Move all constructions and methods from logger to writer
