+++ LoggingConsoleTesting/Program.cs
@@ -3,13 +3,13 @@
 namespace LoggingConsoleTesting
 {
     using Logging;
-    using Logging.Loggers;
+    using Logging.Writers;
 
     internal class Program
     {
         private static void Main()
         {
-            ConsoleLogger logger = new ConsoleLogger();
+            ConsoleWriter logger = new ConsoleWriter();
             logger.WriteMessage("Hola Test Logging");
             logger.MessageLevelProperties.SetPrefix(MessageLevel.Info, "INFORMACION");
             logger.MessageLevelProperties.SetConsoleColor(MessageLevel.Info, ConsoleColor.DarkCyan);
@@ -27,17 +27,17 @@ namespace LoggingConsoleTesting
             Console.WriteLine("Error Messages: " + logger.MessageLevelProperties.GetCounter(MessageLevel.Error));
             Console.WriteLine("Debug Messages: " + logger.MessageLevelProperties.GetCounter(MessageLevel.Debug));
 
-            DebugLogger debugLogger = new DebugLogger();
+            DebuggerWriter debugLogger = new DebuggerWriter();
             debugLogger.WriteMessage("TEST", MessageLevel.Debug);
 
-            FileLogger textLogger = new FileLogger(@"C:\temp", "PruebaLog", WritingMode.Appending);
+            TextFileWriter textLogger = new TextFileWriter(@"C:\temp", "PruebaLog", WritingMode.Appending);
             textLogger.MessageProperties.TimeStampFormat = "hh:mm:ss";
             textLogger.WriteMessage("Test");
             textLogger.WriteMessage("Test2");
             textLogger.WriteMessage("Test");
             textLogger.WriteMessage("Test2");
 
-            FileLogger textLogger2 = new FileLogger(@"C:\temp", "PruebaLog2", WritingMode.Recreating);
+            TextFileWriter textLogger2 = new TextFileWriter(@"C:\temp", "PruebaLog2", WritingMode.Recreating);
             textLogger2.WriteMessage("Test3");
             textLogger2.WriteMessage("Test5");
             textLogger2.WriteMessage("Test4");
